requirements.txt
crewai
openai
python-dotenv
youtube-transcript-api